core: add paging integrity checks
Measure pages that are about to be swapped out by the kernel
to make sure the content is unchanged while in the swap.

Current code only allows swapping out clean pages as we try to
prevent actual guest data from leaking. This is where things
get a bit more complicated as the linux mm notifier chain does
not fully understand that the hyp may not obey the unmap
request when it sees a dirty page. The MM notifier chain may
also give us large guest blobs to unmap on one go which we do
not currently support as we need to verify the pte dirty state
per page basis.

To help us with the task also turn on hardware dirty management
when the hardware supports it. If it doesn't, do 'maybe dirty'
tracking based on the 'writable' nature of the page.

Signed-off-by: Janne Karhunen <Janne.Karhunen@gmail.com>