core: support large guest stage2 maps/unmaps
Previously we have mandated page-size stage2 mappings for
the guest. Add support for all mapping sizes and move the
mapping code to the guest.c.

The memory integrity verification is adapted accordingly.
We can now store the integrity information as well for all
mapping sizes.

Also added documentation on the planned deployment model
in the code.

Signed-off-by: Janne Karhunen <Janne.Karhunen@gmail.com>