core: fix guest pgd allocation accounting when HOST_VMID != 0
PGD is allocated at the time when a guest vmid is not yet known,
so we have attributed the pgd allocation to the host via vmid 0.
However, this does not hold on newer QCOM platforms as the host
vmid can be anything between 0..20 and the zero being something
entirely different.

Signed-off-by: Janne Karhunen <Janne.Karhunen@gmail.com>