core: add support for shared mappings
Previously we have implemented the host blinding via a hack
as the qemu process can still map back limited set of host
pages required for virtio processing. However, while we
limit the set of pages that can be mapped back to the host
and who can actually perform it, the solution is unoptimal
and exploitable.

Start building an infrastructure that mimicks the SEV/TDX
behavior where PTEs can be marked with magical SHARED bit
indicating that the mapping should not be removed from the
host. The kernel patch needs to be adapted accordingly for
this to function, so set_memory_{enc,dec} needs to be
trapped for arm64 as well.

Signed-off-by: Janne Karhunen <Janne.Karhunen@gmail.com>