core: create a proper kvm callback table
Add code to learn the exact kernel entrypoints from the very first
VM bootup. The table is locked on first PSCI_VERSION issued by the
booting VM and made immutable ever since. This should be OK, since
the critical guest bootups are supposed to happen when the host is
still trusted.

Signed-off-by: Janne Karhunen <Janne.Karhunen@gmail.com>