core: implement tdx/sev style memory protection
Add support for 'patched' guests that inform us about the
pages that need to be open for the host communications. In
this model the arm64 guests are expanded to support 'memory
encryption'. Sample guest patch is provided under patches/.

The code for the old model where the guests can be unpatched
is left in place, but it is somewhat less secure. In this
model if the attacker can trick the qemu to touch guest pages
they want to read/write, the pages are automatically mapped
back in if they are in the whitelisted memory sections. There
is a trick to tighten this even further though: if the guests
boot while the host is still part of the TCB we can lock the
virtio areas after the guest kernel initialized.

Signed-off-by: Janne Karhunen <Janne.Karhunen@gmail.com>