core: break-before-make smp support for exception vector
During a break-before-make procedure for a large memory area it is
possible that other cores accesses the memory area for which the
mapping is being altered. This change makes those cores that come
in with data or instruction abort to wait (spinlock) for the mapping
operation to finish. Once the mapping is ready the core is released
back to where the exception was taken from. Core will then be able to
access the area that aborted - or - it will abort again and the abort
will be handled.

Signed-off-by: Jani Hyvonen <jani.hyvonen@digital14.com>