core: contiguous block range tear down support
Arm specification allows marking a set of adjacent translation
table entries as contiguous. This feature allows indicating
contiguous output address ranges to the MMU. This will reduce
the table walk overhead and TLB usage.

When a block belonging to a contiguous range is split into smaller
blocks or pages, the contiguous range needs to be teared as ARM
specifies explicitly the alignment size of the contigous range (in
number of translation table entries) for each granule size.

This patch introduces logic for tearing down existing contiguous
ranges.

Signed-off-by: Jani Hyvonen <jani.hyvonen@digital14.com>