core: stage2 memory range prepare support
HYP_HOST_PREPARE_STAGE2 prepares a range of memory with an existing
stage2 translation table. HYP_HOST_PREPARE_STAGE2 does not change
the memory attributes as normal stage2 mapping operation may do, but
instead it only tears the possible contiguous areas that interleave
the range to be prepared. If the prepared area boundaries interleave
with existing block mappings the block will be split to align with the
mapped area.

If you don't see the use for the API, don't use it. The primary use is
to avoid issues with a centralized TCU during the system runtime when
the mappings change.

HYP_HOST_PREPARE_STAGE2 can be called with similar parameters as
HYP_HOST_MAP_STAGE2.

Signed-off-by: Jani Hyvonen <jani.hyvonen@digital14.com>