core: preliminary support for locking kernel mappings
While we already protect EL1 memory from being modified as mapped,
we don't protect it from being remapped. Implement a base plumbing
for such protections.

When the kernel lock operation is executed for given area we will
find all the kernel page tables from the kernel memory that
describe that area. Each table is then locked as read-only for the
kernel.

Signed-off-by: Janne Karhunen <Janne.Karhunen@gmail.com>