core: hide guest vttbr from the host kernel
The guest vttbr (VMID and BADRR) information is moved entirely to EL2 code
in order to hide it from potentially malicious host kernel.

In addition, guests are now referenced by an vmid-indexed array and
the virt host machine VMID is set to 1 in order to not clash with the
KVM VM initialization phase prior to the VMID assignment.

Signed-off-by: Jani Hyvonen <jani.hyvonen@digital14.com>