core: add encrypted swap support
Encrypt and/or decrypt all the guest pages going in and out of the
system swapfile. We'll do our best tracking writable or dirty guest
pages such that we'll always do minimal amount of work. We use AES
CTR with added integrity verification for the encryption.

As a bonus, add all page protection bits into the page integrity
measurement to make sure that the remap can never alter the initial
page permissions.

Signed-off-by: Janne Karhunen <Janne.Karhunen@gmail.com>