core: reboot and shutdown optimization
Unmap the whole guest stage 2 if the requested range covers the entire guest
RAM area. This function can be used to optimize guest reboot and shutdown
times. Function requires the guest RAM to be contiguous IPA range - return
error if this is not the case and let the default logic to handle the unmap
request.

Signed-off-by: Jani Hyvonen <jani.hyvonen@digital14.com>