core: add support for clean userspace crashes
If and when qemu touches something in the guest it should not have,
make it dump core rather than killing the entire host.

QEMU core dumps on release software generate a valid dump, but the
data where the guest data was is filled with zeroes. All the shared
communication paths remain readable. Debug builds open up all the
guest data.

Signed-off-by: Janne Karhunen <Janne.Karhunen@gmail.com>