core: workaround - do not zero __hyp_running_vcpu at guest_exit
The core specific information on running vcpu is used in __sysreg_restore_el1_state
where the running vcpu pointer is expected to be set when coming out from the guest
and restoring host system registers. If the running vcpu pointer is not set AND the
ARM64_WORKAROUND_SPECULATIVE_AT is set the host context is not restored to SCTLR_EL1
and TCR_EL1.

If ARM64_WORKAROUND_SPECULATIVE_AT is not set this issue is not visible.

Signed-off-by: Jani Hyvonen <jani.hyvonen@digital14.com>