core: bugfixes for guest validate range
Use the incremented physical address for owner check.
Remove the shared area check: 1. the is_any_share as it is currently
implemented does not work with physical addresses. 2. shared check is
unnecessary at this point where the ownership of the address has
already been validated. In ok case address belongs either to the guest
against which the address is validated to or it belongs to the host.
-EPERM is returned if address is already mapped to another guest (it'd
be mapped also if it was a shared address belonging to another guest).

Signed-off-by: Jani Hyvonen <jani.hyvonen@digital14.com>