core: add support for blinded guest debugging
If the hyp is compiled with GUESTDEBUG=1 on the make command line
blinded guests can now be debugged and penetration tested with gdb.

You can stop the guest and the right place will be shown. You can
try to read and write around the guest memory and since the request
to read/write the guest memory comes from the host qemu, you will
see if the hyp would have allowed the access.

Only the gp registers are displayed correctly and stepping inside
the guest will not work. Breakpoints, interrupts and debugger
memory reads are implemented.

Signed-off-by: Janne Karhunen <Janne.Karhunen@gmail.com>