core: support for guest smc and tee functionality
Initial implementation for guest TEE support.

This initial example implementation gives the first VM in the
system the capability to forward SMC calls and have access to
TEE shared memory.

Later on a more elegant implementation to identify the guest
that has SMC capability may be needed.

A hook is added to enable platform implementation to screen SMCs
originated from guest.

Signed-off-by: Jani Hyvonen <jani.hyvonen@unikie.com>