core: remove guest specific el2 translation tables
Reserve all VM runtime EL2 mappings from same table memory pool.
This simplifies the implementation and makes the memory footprint
of a running VM a little bit smaller.

Additionally the implementation has automated cleaning of a guest
EL2 mappings when the guest is being destroyed. This works fine
with one VM but can not be used when there is several VMs running
in the system.

Signed-off-by: Jani Hyvonen <jani.hyvonen@digital14.com>
Change-Id: I7b5e2a5081bedaf6aed67e871a198e53d2fcbfe4