scripts: target-qemu: report umount failure on /proc and /dev (#49)
Rarely umount of /proc and /dev fails, but the cleanup just
passes through without reporting the error. This may lead
a user can do unwanted deletion of these mounted directories
while they clean up the ubuntu directory. Or when users try
to make a tar ball of ubuntu, /proc/kcore may be included
and huge size of tar ball may be created.

This patch tries to check mount status and to report errors
of umount to users, and additionally provides
'make target-qemu-clean' for cleanup of oss/ubuntu/build and
'make target-qemu-distclean' for cleanup of oss/ubuntu.

Change-Id: Ib3cfafc8eaec8c3e86c2816b379d1d2041b21359
Signed-off-by: Sahara <keun-o.park@digital14.com>

Signed-off-by: Sahara <keun-o.park@digital14.com>