core: init: save host context early on
Host context doesn't change, so save it early and use the same context
for every guest <> host switch. This improves the security somewhat,
removes a few reads from the context switch as well as it resolves a
bug where properly timed tlb flush from KVM can crash the host as we
attempt to switch to host when the context had not beed saved yet.

Also add locks for guest_index[] element fetches and sets.

Signed-off-by: Janne Karhunen <Janne.Karhunen@gmail.com>