tools: add support for the 'poor mans profiler'
Virt now supports simple profiling via GDB snapshotting. GDB is
waken up periodically to snapshot the host state via signals and
it seems that we can capture meaningful amount of traces to
produce useful profiling information on where the system spends
most of its time.

scripts/poorman.sh has tunables for the amount of samples that
are being collected and the interval of the samples.

Usage:
make POORMAN=1 run to execute a session, and
make poorman to run analysis.

Signed-off-by: Janne Karhunen <Janne.Karhunen@gmail.com>