kvms_rs: FIFO ticket spinlock implementation in rust
The basic idea and implementation is primarily based on
Nick Piggin's ticket spinlock patch and ARM's ticket spinlock
assembly code. In the Nick Piggin's patch, he claimed that
"On an 8 core (2 socket) Opteron, spinlock unfairness is extremely
noticable, with a userspace test having a difference of up to
2x runtime per thread, and some threads are starved or "unfairly"
granted the lock up to 1 000 000 (!) times. After this patch,
all threads appear to finish at exactly the same time."

As Rust's inline assembly rule is quite strict and not smart unlike
C toolchains(GCC or LLVM), even though rustc is backed by the same LLVM.
So need to be aware that dereferencing the address of a spinlock
should be clearly defined and the dereferenced pointer should be
handed over to the inline assembly. This is the reason why the
core::ptr::addr_of!() is used. Without following this, the compiler
will do a weird optimzation and eventually produce a gabage code.

This is a correct output of ticket spinlock with this patch.
Dump of assembler code for function spin_lock:
   0x00000000c0038b80 <+0>:     mov     w8, wzr
   0x00000000c0038b84 <+4>:     mov     x9, x0
   0x00000000c0038b88 <+8>:     mov     w10, wzr
   0x00000000c0038b8c <+12>:    prfm    pstl1strm, [x0]
   0x00000000c0038b90 <+16>:    ldaxr   w8, [x0]
   0x00000000c0038b94 <+20>:    add     w10, w8, #0x10, lsl #12
   0x00000000c0038b98 <+24>:    stxr    w11, w10, [x0]
   0x00000000c0038b9c <+28>:    cbnz    w11, 0xc0038b90 <spin_lock+16>
   0x00000000c0038ba0 <+32>:    eor     w10, w8, w8, ror #16
   0x00000000c0038ba4 <+36>:    cbz     w10, 0xc0038bbc <spin_lock+60>
   0x00000000c0038ba8 <+40>:    sevl
   0x00000000c0038bac <+44>:    wfe
   0x00000000c0038bb0 <+48>:    ldaxrh  w11, [x9]
   0x00000000c0038bb4 <+52>:    eor     w10, w11, w8, lsr #16
   0x00000000c0038bb8 <+56>:    cbnz    w10, 0xc0038bac <spin_lock+44>
   0x00000000c0038bbc <+60>:    ret

Dump of assembler code for function spin_unlock:
   0x00000000c0038bc0 <+0>:     mov     x8, xzr
   0x00000000c0038bc4 <+4>:     ldrh    w8, [x0]
   0x00000000c0038bc8 <+8>:     add     w8, w8, #0x1
   0x00000000c0038bcc <+12>:    stlrh   w8, [x0]
   0x00000000c0038bd0 <+16>:    ret

Linked-to: https://lwn.net/Articles/267968/
Linked-to: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=314cdbefd1fd0a7acf3780e9628465b77ea6a836
Signed-off-by: Sahara <keun-o.park@digital14.com>