guest support: add support for external memory allocator
Sometimes it's desirable to have a simple memory access control
policy denying all guest memory access from the host. To make that
work with virtio, introduce a new hot-pluggable RAM blob that the
guest can use for talking with the host system.

When the hotplugged ram comes online, the guest can start to use
a new kmalloc flag 'GFP_EXT' for allocating memory from the external
memory pool. This RAM is directly accessible for both ends of the
virtio pipeline. The host will not need any other access to the
guest.

Signed-off-by: Janne Karhunen <Janne.Karhunen@gmail.com>