core: add preliminary support for guest external virtio memory
Sometimes the guest can be configured to allocate virtio memory
from outside of the guest ram. Add support for 100% blinded guests
where the host has zero visibility into the guest ram.

In order to test it, apply the 'ext virtio' patch into the guest
kernel and compile the hyp with USE_EXT_SHARES=1. Also remember
to mark the memory in the guest device tree blob and update the
guestconfig.h accordingly.

Signed-off-by: Janne Karhunen <Janne.Karhunen@gmail.com>