virtiofs: enable virtiofs on guest kernel and install necessary packages
Import a patch for guest kernel which enables virtiofs driver
and dependent configurations needed. And, on host and guest
additionally extra packages will be installed. Some packages
are for host and some are for both host and guest.
As enabling configs and installing packages are harmless,
these will be done by default seamlessly.

Signed-off-by: Sahara <keun-o.park@digital14.com>